Split the pure half of editing out of Database.Edit
Turning a changed activity set back into a Database is pure, and nothing
about it needs the manager, the registry or the disk. It sat in
Database.Edit anyway, which imports Database.Manager, so no module below
the effectful editor could apply an edit.

Database.Rebuild now holds the rebuild and the insert/replace/delete
primitives; Database.Edit keeps everything effectful and imports them.
Pure move, no behaviour change.

The load path needs to apply recorded edits without importing the
effectful editor, which is what the next commit builds on.
