+++ src/Database/Edit.hs
@@ -17,16 +17,10 @@ nothing is mutable. The only fresh allocation is the solver (it holds an
 let the two registry entries share a factorization cache, so the copy gets
 its own.
 
-DELETE is reconstruction, not mutation either. 'deleteActivities' drops a set
-of activities and rebuilds every dependent structure (interning tables,
-indexes, sparse matrices, product index) from the surviving activities. The
-rebuild reuses the exact pure builders that back a freshly-loaded database
-('buildInterningTables' / 'buildTechTriples' / 'buildBioTriples' /
-'buildProductIndex' / 'buildIndexesWithProcessIds'), so a deleted-from
-database is byte-for-byte indistinguishable from one that never carried the
-removed rows. Exchanges in surviving activities that referenced a deleted
-activity are UNLINKED (their activity link reset to nil), leaving the value
-ready for relinking — never silently dropped.
+What an edit does to the activity set itself is pure and lives in
+"Database.Rebuild"; this module is the effectful half around it — which
+database, under what reservation, written where, swapped into the registry
+how.
 
 Memory cost (copy): the copy keeps the source 'Database' alive for as long as
 it is loaded. Structural sharing means we don't re-allocate the activity/flow
@@ -35,15 +29,11 @@ while any copy of it is loaded.
 -}
 module Database.Edit (
     copyDatabase,
-    deleteActivities,
-    deleteActivitiesWith,
     resolveDeleteSelection,
     DeleteSelection (..),
     DeleteRequest (..),
     DeleteOutcome (..),
     deleteActivitiesInDB,
-    insertActivities,
-    replaceActivities,
     MutationOutcome (..),
     mutateUploadedDatabase,
     WriteVerb (..),
@@ -76,12 +66,7 @@ import System.Directory (
 import System.FilePath (makeRelative, splitDirectories, takeDirectory, (</>))
 
 import Config (DatabaseConfig (..))
-import Database (
-    applyStructuredFilters,
-    buildIndexesWithProcessIds,
-    buildProductIndex,
-    findActivitiesByFields,
- )
+import Database (applyStructuredFilters, findActivitiesByFields)
 import Database.Author (
     AuthorContext (..),
     AuthoredActivity,
@@ -100,13 +85,11 @@ import Database.Manager (
     getMergedUnitConfig,
     relinkDatabase,
  )
-import Database.MatrixBuild (
-    InterningTables (..),
-    buildBioTriples,
-    buildInterningTables,
-    buildSupplierRefUnits,
-    buildTechTriples,
-    collectBioFlowOrder,
+import Database.Rebuild (
+    deleteActivitiesWith,
+    insertActivities,
+    renderKey,
+    replaceActivities,
  )
 import Database.Upload (DatabaseFormat (..), slugify)
 import qualified Database.UploadedDatabase as UploadedDB
@@ -115,20 +98,14 @@ import qualified Search.BM25 as BM25
 import Service (bm25Retrieve)
 import SharedSolver (createSharedSolver)
 import Types (
-    Activity (..),
-    BiosphereFlow (..),
     Database (..),
-    Exchange (..),
     ProcessId,
     SparseTriple (..),
-    TechnosphereFlow (..),
-    UUID,
     findProcessId,
     findProcessIdByActivityUUID,
     initializeRuntimeFields,
     parseUUIDPair,
  )
-import UnitConversion (UnitConfig, defaultUnitConfig)
 
 {- | Copy a loaded database into the runtime registry under the slugified
 @newName@.
@@ -215,222 +192,6 @@ renameConfig newName cfg =
         , dcIsUploaded = True
         , dcDeletable = True
         }
-
--- ---------------------------------------------------------------------------
--- Delete by selection
--- ---------------------------------------------------------------------------
-
-{- | Remove a set of activities (by 'ProcessId') and rebuild every dependent
-structure. Uses 'defaultUnitConfig'; effectful call sites that carry a merged
-unit config (the only place where non-SI conversions matter) should call
-'deleteActivitiesWith' instead so a coefficient is never silently dropped on
-an unknown-unit conversion.
--}
-deleteActivities :: [ProcessId] -> Database -> Either Text Database
-deleteActivities = deleteActivitiesWith defaultUnitConfig
-
-{- | 'deleteActivities' with an explicit 'UnitConfig'.
-
-Steps, all pure:
-
-  1. Resolve the delete set to the @(activityUUID, productUUID)@ keys it
-     occupies in 'dbProcessIdTable', validating that every requested
-     'ProcessId' exists (no silent skip).
-  2. Drop those keys; for every surviving activity, UNLINK any technosphere /
-     waste exchange whose @(activityLink, flow)@ pointed at a deleted key —
-     reset its activity link to 'UUID.nil' and clear the stale process link.
-  3. Rebuild interning tables, indexes, matrices and the product index from
-     the surviving activity map via the shared loader builders.
-
-Fails (Left) when an out-of-range 'ProcessId' is requested, when the result
-would be empty, or when matrix construction reports an inconsistency (e.g. an
-unknown unit conversion under the given config).
--}
-deleteActivitiesWith :: UnitConfig -> [ProcessId] -> Database -> Either Text Database
-deleteActivitiesWith unitConfig pids db = do
-    deletedKeys <- resolveDeleteKeys db pids
-    let survivors = surviving db deletedKeys
-        survivingKeys = M.keysSet survivors
-        unlinkedMap = M.map (unlinkActivity survivingKeys) survivors
-    if M.null unlinkedMap
-        then Left "Refusing to delete: the result would have no activities"
-        else rebuildFromActivities unitConfig db unlinkedMap
-
-{- | Resolve each requested 'ProcessId' to its @(activityUUID, productUUID)@ key.
-Every id must be in range — an out-of-range id is a caller error, surfaced as
-'Left' rather than silently ignored. The result is a 'Set' so membership tests
-during unlinking are @O(log n)@.
--}
-resolveDeleteKeys :: Database -> [ProcessId] -> Either Text (S.Set (UUID, UUID))
-resolveDeleteKeys db = fmap S.fromList . traverse keyOf
-  where
-    table = dbProcessIdTable db
-    n = V.length table
-    keyOf pid
-        | i >= 0 && i < n = Right (table V.! i)
-        | otherwise = Left ("Delete: ProcessId out of range: " <> T.pack (show pid))
-      where
-        i = fromIntegral pid
-
--- | The activity map keyed by @(activityUUID, productUUID)@, minus the deleted keys.
-surviving :: Database -> S.Set (UUID, UUID) -> M.Map (UUID, UUID) Activity
-surviving db deletedKeys =
-    M.fromList
-        [ (key, dbActivities db V.! i)
-        | i <- [0 .. V.length (dbActivities db) - 1]
-        , let key = dbProcessIdTable db V.! i
-        , not (S.member key deletedKeys)
-        ]
-
-{- | Reset technosphere / waste exchanges on a surviving activity whose producer
-link no longer resolves to a surviving @(activityUUID, productUUID)@ key.
-
-The link target is exactly that pair: 'findProducer' resolves
-@(activityLink, flowId)@ against the interning table, so a multi-product
-activity that keeps at least one product stays a valid target for exchanges
-pointing at a *surviving* product, while exchanges pointing at a deleted
-product are unlinked. A biosphere exchange has no producer link and is
-returned unchanged. The stale 'ProcessId' link is always cleared because
-deletion renumbers every 'ProcessId'. An already-orphan link
-(@activityLink == nil@) stays orphan.
--}
-unlinkActivity :: S.Set (UUID, UUID) -> Activity -> Activity
-unlinkActivity survivingKeys act =
-    act{exchanges = map unlinkExchange (exchanges act)}
-  where
-    dangling link flow = link /= UUID.nil && not (S.member (link, flow) survivingKeys)
-    unlinkExchange ex = case ex of
-        BiosphereExchange{} -> ex
-        TechnosphereExchange{techActivityLinkId = link, techFlowId = flow}
-            | dangling link flow ->
-                ex{techActivityLinkId = UUID.nil, techProcessLinkId = Nothing}
-            | otherwise -> ex{techProcessLinkId = Nothing}
-        WasteExchange{waActivityLinkId = link, waFlowId = flow}
-            | dangling link flow ->
-                ex{waActivityLinkId = UUID.nil, waProcessLinkId = Nothing}
-            | otherwise -> ex{waProcessLinkId = Nothing}
-
-{- | Rebuild a 'Database' from a surviving activity map, reusing the exact pure
-builders that back a freshly-loaded database. Flow / unit tables are carried
-over unchanged: deletion removes activities, never the flow or unit
-vocabulary, so a relink can still resolve the surviving links. Runtime fields
-(synonym, flow-name, BM25 indexes) are reset to their unloaded state; the
-effectful caller re-attaches them with the merged synonym DB.
--}
-rebuildFromActivities :: UnitConfig -> Database -> M.Map (UUID, UUID) Activity -> Either Text Database
-rebuildFromActivities unitConfig db activityMap =
-    let tables = buildInterningTables activityMap
-        supplierRefUnits = buildSupplierRefUnits (dbUnits db) (itActivities tables)
-        indexes = buildIndexesWithProcessIds (itActivities tables) (itProcessIdTable tables)
-        bioFlowUUIDs = collectBioFlowOrder (itActivities tables)
-        bioTriples = buildBioTriples bioFlowUUIDs tables
-        productIndex = buildProductIndex (itActivities tables) (itProcessIdTable tables) (dbTechFlows db)
-     in case buildTechTriples unitConfig (dbUnits db) tables supplierRefUnits of
-            Left err -> Left err
-            Right (techTriples, _warnings) ->
-                Right
-                    db
-                        { dbProcessIdTable = itProcessIdTable tables
-                        , dbProcessIdLookup = itProcessIdLookup tables
-                        , dbActivityUUIDIndex = itActivityUUIDIndex tables
-                        , dbActivityProductsIndex = itActivityProductsIndex tables
-                        , dbProductIndex = productIndex
-                        , dbActivities = itActivities tables
-                        , dbIndexes = indexes
-                        , dbTechnosphereTriples = techTriples
-                        , dbBiosphereTriples = bioTriples
-                        , dbActivityIndex = V.generate (fromIntegral (itActivityCount tables)) fromIntegral
-                        , dbBiosphereOrder = bioFlowUUIDs
-                        , dbActivityCount = itActivityCount tables
-                        , dbBiosphereCount = fromIntegral (V.length bioFlowUUIDs)
-                        , -- Cross-DB links may now reference deleted activities; clear so a
-                          -- subsequent relink rebuilds them against the surviving set.
-                          dbCrossDBLinks = []
-                        , -- Linking stats describe the pre-delete set (input count, completeness,
-                          -- missing suppliers); reset to the fresh-load default so the setup/status
-                          -- endpoint never reports stale numbers until the next relink.
-                          dbLinkingStats = mempty
-                        , -- Runtime-only indexes are reset; re-attached by the effectful caller.
-                          dbSynonymDB = Nothing
-                        , dbFlowsByName = M.empty
-                        , dbFlowsByCAS = M.empty
-                        , dbProductSearchIndex = M.empty
-                        , dbBM25Index = Nothing
-                        }
-
--- ---------------------------------------------------------------------------
--- Insert / replace
--- ---------------------------------------------------------------------------
-
-{- | Whether a batch means to add rows the database does not have, or to
-rewrite rows it does. There is deliberately no upsert: a mistyped identity
-must fail, not quietly become a second copy of an activity the author thought
-they were correcting.
--}
-data WriteIntent = Insert | Replace
-
-{- | Add authored activities to a database and rebuild everything that depends
-on them.
-
-Every key must be absent — 'Database.Author' mints identity from the activity's
-own name, location, product and unit, so a key that already exists means the
-author is re-describing something the database already holds and wants
-'replaceActivities' instead.
--}
-insertActivities :: UnitConfig -> [ResolvedInsert] -> Database -> Either Text Database
-insertActivities = applyResolved Insert
-
-{- | Rewrite activities the database already holds, keeping their identity.
-
-Every key must be present. Replacing keeps the old version's flows in the
-vocabulary, exactly as deletion does: a flow no activity uses any more is
-inert, and dropping it would break a relink that still resolves through it.
--}
-replaceActivities :: UnitConfig -> [ResolvedInsert] -> Database -> Either Text Database
-replaceActivities = applyResolved Replace
-
-{- | Shared write path. The vocabulary the batch brings (its product flows and
-any new biosphere flows) lands in the same step as the activities that
-reference it, so no intermediate value exists in which an exchange points at a
-flow the database cannot name.
-
-An empty batch returns the database untouched rather than rebuilding: a
-rebuild resets cross-database links ('rebuildFromActivities'), so treating "no
-activities to write" as a no-op is what keeps a caller's empty request from
-silently unlinking the database.
--}
-applyResolved :: WriteIntent -> UnitConfig -> [ResolvedInsert] -> Database -> Either Text Database
-applyResolved intent unitConfig inserts db
-    | null inserts = Right db
-    | otherwise = do
-        let existing = surviving db S.empty
-        checkKeys intent existing (map riKey inserts)
-        rebuildFromActivities
-            unitConfig
-            db
-                { dbTechFlows = M.union (indexBy tfId (concatMap riNewTechFlows inserts)) (dbTechFlows db)
-                , dbBioFlows = M.union (indexBy bfId (concatMap riNewBioFlows inserts)) (dbBioFlows db)
-                }
-            (M.union (M.fromList [(riKey i, riActivity i) | i <- inserts]) existing)
-  where
-    indexBy key xs = M.fromList [(key x, x) | x <- xs]
-
-{- | Refuse a batch whose keys contradict the intent, naming every offending
-identity at once so a long batch is not fixed one round-trip at a time.
--}
-checkKeys :: WriteIntent -> M.Map (UUID, UUID) Activity -> [(UUID, UUID)] -> Either Text ()
-checkKeys intent existing keys = case intent of
-    Insert -> refuse "already exists" (filter (`M.member` existing) keys)
-    Replace -> refuse "does not exist" (filter (not . (`M.member` existing)) keys)
-  where
-    refuse _ [] = Right ()
-    refuse what offenders =
-        Left $
-            "Refusing to write: "
-                <> T.intercalate ", " (map renderKey offenders)
-                <> " "
-                <> what
-                <> " in this database"
 
 -- ---------------------------------------------------------------------------
 -- Writing authored activities
@@ -587,10 +348,6 @@ loadedDependencies :: DatabaseManager -> Database -> IO [Database]
 loadedDependencies manager db = do
     loadedDbs <- readTVarIO (dmLoadedDbs manager)
     pure [ldDatabase ld | name <- dbDependsOn db, Just ld <- [M.lookup name loadedDbs]]
-
--- | @activityUUID_productUUID@, the identity a process is addressed by.
-renderKey :: (UUID, UUID) -> Text
-renderKey (actUUID, prodUUID) = UUID.toText actUUID <> "_" <> UUID.toText prodUUID
 
 {- | Plain-language rendering, for callers with nowhere to put a status code.
 An HTTP caller maps the constructors to codes instead.
